Add global system prompt and refactor config types
基本完善用户配置面板基础功能实现，细节待完善；提供单文件类型支持；优化各种逻辑

+++ src-tauri/src/completion/mod.rs
@@ -21,7 +21,7 @@ pub struct BalanceMessage {
 #[derive(serde::Serialize, serde::Deserialize, Clone)]
 pub struct UserConfig {
     pub temperature: f32,
-    pub max_tokens: f32,
+    pub max_tokens: u32,
     pub top_p: f32,
     pub frequency_penalty: f32,
 }
